requirements.txt
# Web Framework
fastapi
uvicorn

# MongoDB
motor

# Authentication
python-jose
passlib
bcrypt
python-multipart  # For OAuth2 form handling and file uploads

# System monitoring
psutil

# Data Validation
pydantic
email-validator  # Required for pydantic[email]

# HTTP Client
httpx

# Google OAuth and Maps
google-auth
google-auth-oauthlib
google-auth-httplib2

# Environment Variables
python-dotenv

# Utilities and Date/Time
pytz
python-dateutil

# Email Templates
Jinja2

# File handling
python-magic  # For file type detection
Pillow  # For image processing

# Payment processing
stripe
paypalrestsdk

# Push notifications
pyfcm

# Rate limiting
slowapi
redis  # For rate limiting backend

# Additional utilities
uuid
qrcode  # For generating QR codes
geopy  # For location services

pydantic_settings==2.9.1